% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_li6800.R
\name{read_li6800}
\alias{read_li6800}
\title{Title read LI-6800 excel data files.}
\usage{
read_li6800(data_files, startRow = 15, add_tags = TRUE, tags = NULL, ...)
}
\arguments{
\item{data_files}{path of the LI-6800 data files}

\item{startRow}{where the header starts, ie the row start with `obs`}

\item{add_tags}{whether to add a column that contains a tag, it is ued for batch reading of excel files,
default to be true}

\item{tags}{when add_tags is true, use a characters as a tag,
if NULL, use file name as defaut tags.}

\item{...}{other paras from read_excel}
}
\value{
a data frame of tidy data
}
\description{
Title read LI-6800 excel data files.
}
\details{
note the excel must select Options ---> formula --> Enable iterative calculation,or it will be filled
in reading cells that contain formulas
}
\examples{
\dontrun{
read_li6800('data.xlsx')
}
}
