% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_li6800_folder.R
\name{read_li6800_folder}
\alias{read_li6800_folder}
\title{Title read LI-6800 excel data files in batch.}
\usage{
read_li6800_folder(
  data_folder = NULL,
  add_tags = TRUE,
  tags = NULL,
  col_names = FALSE,
  ...
)
}
\arguments{
\item{data_folder}{path of tthe folder that contains LI-6800 excel data files}

\item{add_tags}{whether to add a column that contains a tag, it is ued for batch reading of excel files,
default to be true}

\item{tags}{when add_tags is true, use a characters as a tag,
if NULL, use file name as defaut tags.}

\item{col_names}{must be false to make sure the data import}

\item{...}{other paras from read_excel}
}
\value{
a data frame of tidy data
}
\description{
Title read LI-6800 excel data files in batch.
}
\details{
note the excel must select Options ---> formula ---> Enable iterative calculation,or it will be filled
in reading cells that contain formulas
}
\examples{
\dontrun{
read_li6800_folder('./data')
}
}
